Made something happen
added the JFrame to main, made it show up, created default constructors
and some methods for Answer, Question, and Person. A person can be added
to the JFrame as I have done and that, for now, will add a randomly
selected question (random start "points", and random selection from that
point level) and put it on, nothing happens yet if you click buttons,
that will be done later.